feat: remove focus from form toggle buttons
